gaurdrails/requirements.txt
torch
transformers
pyyaml
faiss-cpu
azure-ai-inference
azure-core
langgraph
aiohttp>=3.10.11
fastapi>=0.103.0
fastembed<0.4.1,>=0.2.2
langchain<0.4.0,>=0.2.14
langchain-community<0.4.0,>=0.0.16
lark<1.2.0,>=1.1.7
nest-asyncio>=1.5.6
onnxruntime<2.0.0,>=1.17.0
onnxruntime<=1.19.2,>=1.17.0
pandas<3,>=1.4.0
prompt-toolkit>=3.0
rich>=13.5.2
simpleeval>=0.9.13
starlette>=0.27.0
typer>=0.8
uvicorn>=0.23
watchdog>=3.0.0
httpx<0.25.0,>=0.24.1
